requirements.txt
# Movo - Python AI Chat Application

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
streamlit==1.30.0
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.5.0
python-dotenv==1.0.0
httpx==0.26.0
sse-starlette==2.0.0